Improved mininet cleanup
Kill only specific processes

+++ start_network.py
@@ -5,16 +5,26 @@ import os
 import sys
 import time
 
-from mininet.clean import cleanup
 from mininet.net import Mininet
 from mininet.topo import LinearTopo
 
 BOOT_PORT = 9000
 
 
-def cleanup_logs():
-    for f in os.listdir(config.LOG_DIR):
-        os.remove(f)
+def cleanup(logs=False):
+
+    print("Killing all nodes\n")
+    os.system("pkill -SIGINT -f '^python3 -u start_node.py'")
+
+    print("Killing all xterms\n")
+    os.system("killall -SIGKILL xterm")
+
+    # Call standard mn cleanup
+    os.system("mn --clean")
+
+    if logs:
+        for f in os.listdir(config.LOG_DIR):
+            os.remove(f)
 
 
 def start_network(nodes=3):
@@ -29,7 +39,7 @@ def start_network(nodes=3):
     # The first node acts as a bootstrapper for other
     boot_ip = hosts[0].IP()
     hosts[0].cmd(
-        'xterm -hold -geometry 130x40+0+900 -title "bootstrap %s %d" -e python3 start_node.py %s %d &' % (
+        'xterm -hold -geometry 130x40+0+900 -title "bootstrap %s %d" -e python3 -u start_node.py %s %d &' % (
             boot_ip, BOOT_PORT, boot_ip, BOOT_PORT)
     )
 
@@ -49,13 +59,8 @@ def start_network(nodes=3):
 
     raw_input('Press enter to stop all nodes.')
 
-    print("Killing all nodes\n")
-    os.system("killall -SIGINT python3")
-
-    print("Killing all xterms\n")
-    os.system("killall -SIGINT xterm")
-
     net.stop()
+
     cleanup()
 
 if __name__ == '__main__':
